Vector3の比較演算子==を定義。また、そこでfloat値の一致判定を行う必要が出てきたが、そのような処理は何度も行う可能性が高いので、そういった共通処理をinternal/commonフォルダ以下で実装するものとし、CommonUtil.h, CommonUtil.cppを実装。

+++ libsources/FbxAgent.cpp
@@ -178,7 +178,7 @@ if (ret != FbxAgentErrorCode::FBX_AGENT_SUCCESS)
         // TODO : UV座標をどういう風に展開すればVulkanで扱いやすい形に出来るのか考える。マッピングモードの違いにどう対処する？
         // 同じFbxMesh内の頂点は全て同じレイヤを持っているという前提を置いている
         int layerCount = mesh->GetLayerCount();
-        int startIndex = vertexUVs.size();
+        int startIndex = (int)vertexUVs.size();
         int vertexCount = mesh->GetPolygonVertexCount();
 
         for (int i = 0; i < vertexCount; i++)
